// src/interfaces/swap.ts
export interface Pool {
  id: number;
  tokenIds: string[];
  supplies: { [key: string]: string };
  shareSupply: string;
  fee: number;
  pool_kind: string;
  tvl?: string;
  partialAmountIn?: string;
  rates?: {
    [id: string]: string;
  };
  Dex?: string;
  pairAdd?: string;
  token0_ref_price?: string;
}
// ...
export interface ReservesMap {
  [index: string]: string;
}
export type SwapContractType =
  | "Ref_Classic"
  | "Ref_DCL"
  | "Orderly"
  | "Trisolaris";
export enum PoolMode {
  PARALLEL = "parallel swap",
  SMART = "smart routing",
  SMART_V2 = "stableSmart",
  STABLE = "stable swap",
}
export enum SWAP_MODE {
  NORMAL = "normal",
  LIMIT = "limit",
}
export interface RoutePool {
  amounts: string[];
  fee: number;
  id: number;
  reserves: ReservesMap;
  shares: string;
  token0_ref_price: string;
  token1Id: string;
  token1Supply: string;
  token2Id: string;
  token2Supply: string;
  updateTime: number;
  partialAmountIn?: string | number | Big;
  gamma_bps?: Big;
  supplies?: ReservesMap;
  tokenIds?: string[];
  x?: string;
  y?: string;
}
export interface EstimateSwapView {
  estimate: string;
  pool: Pool | null;
  intl?: any;
  partialAmountIn?: string;
  dy?: string;
  status?: PoolMode;
  token?: TokenMetadata;
  noFeeAmountOut?: string;
  inputToken?: string;
  outputToken?: string;
  nodeRoute?: string[];
  tokens?: TokenMetadata[];
  routeInputToken?: string;
  routeOutputToken?: string;
  route?: RoutePool[];
  allRoutes?: RoutePool[][];
  allNodeRoutes?: string[][];
  totalInputAmount?: string;
  overallPriceImpact?: string;
  contract?: SwapContractType;
  percent?: string;
}
// ...
export interface IEstimateServerResult {
  estimatesFromServer: IEstimateSwapServerView;
  source: "server";
  tag: string;
  poolsMap: Record<string, any>;
  tokensMap: Record<string, TokenMetadata>;
}
// ...
export interface IEstimateSwapServerView {
  amount_in: string;
  amount_out: string;
  contract_in: string;
  contract_out: string;
  routes: IServerRoute[];
  contract?: string;
}
export interface IServerRoute {
  amount_in: string;
  min_amount_out: string;
  pools: IServerPool[];
  tokens: TokenMetadata[];
}